Finish moving FaqContent under 'home' app
