+++ getyourdata/data_request/admin.py
@@ -3,7 +3,7 @@ from django.contrib import admin
 from modeltranslation.admin import TranslationAdmin
 
 from data_request.models import DataRequest, AuthenticationContent
-from data_request.models import FaqContent, RequestContent
+from data_request.models import RequestContent
 from data_request.models import FeedbackMessageContent
 
 
@@ -15,8 +15,3 @@ class RequestContentAdmin(TranslationAdmin):
 @admin.register(FeedbackMessageContent)
 class FeedbackMessageContentAdmin(TranslationAdmin):
     list_display = ['name']
-
-
-@admin.register(FaqContent)
-class FaqContentAdmin(TranslationAdmin):
-    list_display = ['title', 'priority']
