+++ getyourdata/data_request/translation.py
@@ -1,5 +1,5 @@
 from modeltranslation.translator import register, TranslationOptions
-from data_request.models import FaqContent, RequestContent
+from data_request.models import RequestContent
 from data_request.models import FeedbackMessageContent
 
 
@@ -11,8 +11,3 @@ class RequestContentTranslationOptions(TranslationOptions):
 @register(FeedbackMessageContent)
 class FeedbackMessageContentTranslationOptions(TranslationOptions):
     fields = ('name', 'header', 'pdf_copy', 'footer')
-
-
-@register(FaqContent)
-class FaqContentTranslationOptions(TranslationOptions):
-    fields = ('title', 'content')
