+++ getyourdata/home/translation.py
@@ -1,8 +1,13 @@
 from modeltranslation.translator import register, TranslationOptions
 
-from home.models import HomePage
+from home.models import HomePage, FaqContent
 
 
 @register(HomePage)
 class HomePageTranslationOptions(TranslationOptions):
     fields = ('content',)
+
+
+@register(FaqContent)
+class FaqContentTranslationOptions(TranslationOptions):
+    fields = ('title', 'content')
